Organización de rutas

+++ src/main/java/com/proyecto/veterinaria/Model/Pago.java
@@ -3,6 +3,7 @@ package com.proyecto.veterinaria.Model;
 import jakarta.persistence.*;
 import lombok.*;
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "pago")
@@ -18,7 +19,7 @@ public class Pago {
     private Long idPago;
     private LocalDateTime fechaHora;
     private String metodo;
-    private Double monto; // Usamos Double para el monto, podrías considerar BigDecimal para mayor precisión en aplicaciones financieras.
+    private BigDecimal monto; 
     private String estado; // Ejemplo: "COMPLETADO", "ANULADO"
   //Relacón bidireccional de pago con factura 
     @ManyToOne(fetch = FetchType.LAZY)
